maxo/requirement.txt
asgiref==3.3.1
certifi==2020.12.5
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-crispy-forms==1.10.0
django-formtools==1.0
django-import-export==2.5.0
django-reversion==3.0.8
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
pytz==2020.5
PyYAML==5.3.1
six==1.10.0
sqlparse==0.4.1
tablib==3.0.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
